Test for building instance DAGs
